[style] main 페이지 디자인 개선
