+++ src/App.css
@@ -8,3 +8,7 @@
   padding: 0;
   box-sizing: border-box;
 }
+a {
+  font-style: normal;
+  text-decoration: none;
+}
