+++ src/components/Bookshelf.css
@@ -1,115 +1,85 @@
 .bookshelf-section {
+  width: 120vh; /* 넓이 조정 */
+  height: 100px; /* 높이 조정 */
+  background-color: #f9f9f9;
   position: relative;
-  width: 1000px; /* 넓이 조정 */
-  height: 400px; /* 높이 조정 */
-  margin: 0 auto; /* 중간 정렬y */
-  z-index: 1;
 }
-
-.shelf-img {
-  width: 100%;
-  height: auto;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-}
-
-.bookshelf {
-  display: flex;
-  align-items: flex-end;
-  position: absolute;
-  bottom: 125px; /* 책 위치 조정 */
-  left: 50%;
-  transform: translateX(-50%);
+.bookshelf-section .bookshelf {
+  width: 75%;
+  height: 100%;
   padding-right: 0;
-  z-index: 2; /*우선 순위*/
+  margin-right: -12%;
+  display: flex;
+  position: absolute;
+  right: 0;
+}
+.bookshelf-section .bookshelf .book-container {
+  width: 100%;
+  height: 100%;
+  position: relative;
+  bottom: -50%;
 }
 
-.book {
-  width: 50px;
-  border-radius: 5px;
-  margin: 0 2px;
-  /* transition: transform 0.3s ease; */
+/* 책 */
+.bookshelf-section .bookshelf .book {
+  width: 100%;
+  transform: scaleX(-1);
+  position: absolute;
+  right: 0;
+  bottom: 0;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(1) .book {
+  width: 135%;
+  margin-right: -286%;
+  transition: opacity 1s ease-in-out;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(2) .book {
+  width: 69%;
+  margin-right: -250%;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(3) .book {
+  width: 51%;
+  margin-right: -197%;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(4) .book {
+  width: 49%;
+  margin-right: -143%;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(5) .book {
+  width: 77%;
+  margin-right: -116%;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(6) .book {
+  width: 57%;
+  margin-right: -70%;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(7) .book {
+  width: 110%;
+  margin-right: -70%;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(8) .book {
+  width: 37%;
+  margin-right: -64%;
+}
+.bookshelf-section .bookshelf .book-container:nth-child(9) .book {
+  width: 40%;
+  margin-left: -0.5%;
 }
 
-.pot {
-  width: 80px;
-  margin-left: 60px;
-}
 
-.book:hover {
-  /* transform: translateY(-50px); */
+/* 책 클릭 이벤트 */
+.bookshelf-section .bookshelf .book:hover {
   cursor: pointer;
-  transform: scale(1.4);
-  transition-duration: 0.5s;
 }
-
-@media (max-width: 1200px) {
-  .bookshelf-section {
-    max-width: 800px;
-    max-height: 300px;
-  }
-  .bookshelf {
-    bottom: 100px;
-  }
-
-  .book {
-    width: 40px;
-    border-radius: 5px;
-    margin: 0 2px;
-    /* transition: transform 0.3s ease; */
-  }
-
-  .pot {
-    margin-left: 50px;
-    width: 60px;
-  }
+.bookshelf-section .bookshelf .book-container .book:hover {
+  opacity: 0.5;
+  animation: fadeInOut 2s infinite;
 }
-
-@media (max-width: 768px) {
-  .bookshelf-section {
-    max-width: 600px;
-    max-height: 225px;
+@keyframes fadeInOut {
+  0%, 100% {
+    opacity: 1;
   }
-  .bookshelf {
-    bottom: 75px;
-  }
-
-  .book {
-    width: 26px;
-    border-radius: 5px;
-    margin: 0 2px;
-    /* transition: transform 0.3s ease; */
-  }
-
-  .pot {
-    margin-left: 40px;
-    width: 38px;
-  }
-}
-
-@media (max-width: 480px) {
-  .bookshelf-section {
-    max-width: 300px;
-    max-height: 120px;
-  }
-  .bookshelf {
-    bottom: 38px;
-  }
-  .book {
-    width: 10vw;
-    max-width: 40px;
-  }
-
-  .book {
-    width: 15px;
-    border-radius: 5px;
-    margin: 0 2px;
-    /* transition: transform 0.3s ease; */
-  }
-
-  .pot {
-    margin-left: 10px;
-    width: 20px;
+  50% {
+    opacity: 0.5;
   }
 }
