+++ src/components/Header.css
@@ -28,99 +28,6 @@
   color: #000000;
 }
 
-.search-bar {
-  display: flex;
-  align-items: center;
-  width: 50%;
-  height: 40px;
-  padding: 11px;
-  margin: 0 20px 0 20px;
-  box-shadow: 4px 4px 4px rgba(0, 0, 0, 0.25);
-  border-radius: 30px;
-  position: relative;
-}
-
-.search-input {
-  width: 100%;
-  border: none;
-  outline: none;
-  padding: 0 40px 0 20px;
-  border-radius: 30px;
-  font-size: 16px;
-  color: black;
-}
-
-.search-icon {
-  position: absolute;
-  right: 15px;
-  width: 17.5px;
-  height: 17.5px;
-}
-
-.search-results {
-  position: absolute;
-  top: 50px;
-  left: 0;
-  right: 0;
-  background-color: white;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  max-height: 600px;
-  overflow-y: auto;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  z-index: 10;
-}
-
-.search-result-item {
-  display: flex;
-  align-items: center;
-  padding: 10px;
-  border-bottom: 1px solid #eee;
-}
-
-.search-result-item img {
-  width: 60px;
-  height: 90px;
-  object-fit: cover;
-  margin-right: 10px;
-}
-
-.search-book-info {
-  flex-grow: 1;
-}
-
-.search-book-title {
-  font-weight: bold;
-  font-size: 16px;
-}
-
-.search-book-author,
-.search-book-price,
-.search-book-publisher {
-  font-size: 14px;
-  color: #555;
-}
-
-.buy-button {
-  width: 50px;
-  height: 50px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0; /* 이미지가 줄어들지 않게 하기 */
-}
-
-.buy-button img {
-  width: 100%;
-  height: auto;
-  object-fit: cover;
-}
-
-.error-message {
-  padding: 10px;
-  color: #555;
-}
-
 .menu {
   display: flex;
   align-items: center;
