+++ src/components/Main.jsx
@@ -1,6 +1,7 @@
 import Header from './Header';
 import { Outlet } from 'react-router-dom';
 
+
 function Main() {
   return (
     <>
